etl/requirements.txt
wheel
image
html2text
scrapy-splash
python-dateutil
python-dotenv==0.13.0
Scrapy==1.8.0
requests==2.23.0
vobject==0.9.6.1
xmltodict
overrides==3.1.0